Code with 1 box on face, have defects at some poses but I think it is from the multiple angle think.
Too late and not yet made the food so will start preparing it but I am really happy with the progress, that we both did.

To be honest I would have not reached this clean solution without your very (as much as possible) c;ean angle detection.

Thank you miamor, M7taga Ragel